refactor: move and rename type test
